GraphHistory renamed, split into three
