+++ idtrack/_db.py
@@ -28,5 +28,6 @@ class DB:
     # Protected Non-int Version Strings
     no_old_node_id = "Void"
     no_new_node_id = "Retired"
+    alternative_versions = {no_new_node_id, no_old_node_id}
 
     external_search_settings = {"jump_limit": 2, "synonymous_max_depth": 3, "backbone_node_type": "ensembl_gene"}
